= 1.4.5 = * Minor security improvements * Titan Framework update (1.7.3) * Chart.js library update * Better responsiv support on dashboard charts * Added charts data on hover
git-svn-id: https://plugins.svn.wordpress.org/adblock-notify-by-bweb/trunk@1068081 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ adblock-notify-functions.php
@@ -1,5 +1,13 @@
 <?php
 /***************************************************************
+ * SECURITY : Exit if accessed directly
+***************************************************************/
+if ( !defined( 'ABSPATH' ) ) {
+	die( 'Direct acces not allowed!' );
+}
+
+
+/***************************************************************
  * Insert elements in the DOM : HTML & SCRIPT
  ***************************************************************/
 function an_prepare() {
@@ -281,8 +289,7 @@ function an_reset_stats() {
     if ( $screen->id != 'toplevel_page_'. AN_ID )
         return;
 		
-	if($_GET['an-reset'] == 'true'){
-		
+	if( isset( $_GET[ 'an-reset' ] ) && $_GET[ 'an-reset' ] == 'true' ){
 		delete_option( 'adblocker_notify_counter' );
 		add_action( 'admin_notices', 'an_stats_notice' );
 	}
